typescript対応
