+++ front/index.html
@@ -1,0 +1,2 @@
+<div id="root"></div>
+<script type="module" src="/src/index.tsx"></script>
