+++ front/src/coomponents/pages/CreateRoomPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CreateRoom from "../views/CreateRoom.js"
+import CreateRoom from "../views/CreateRoom.jsx"
 
 export class CreateRoomPage extends React.Component {
   render() {
