+++ front/src/coomponents/pages/HomePage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Home from "../views/Home.js";
+import Home from "../views/Home.jsx";
 
 export class HomePage extends React.Component {
   render() {
