+++ front/src/coomponents/pages/RegisterUserPage.jsx
@@ -1,6 +1,5 @@
 import React from "react";
-import RegisterUser from "../views/RegisterUser.js";
-
+import RegisterUser from "../views/RegisterUser.jsx"
 export class RegisterUserPage extends React.Component {
   render() {
     return <RegisterUser />;
