+++ front/src/coomponents/pages/UserInfoPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import UserInfo from "../views/UserInfo.js";
+import UserInfo from "../views/UserInfo.jsx";
 
 export class UserInfoPage extends React.Component {
   render() {
