+++ front/src/coomponents/templates/InputImage.jsx
