--- front/src/index.js
@@ -1,12 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-// import App from "./App";
-import "./index.css";
-import { Routing } from "./Routing.js";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  // <React.StrictMode>
-    <Routing />,
-  {/* </React.StrictMode> */}
-);
